# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Preloadet",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/post-category": {
            "get": {
                "tags": [
                    "PostCategory"
                ],
                "summary": "All Post Category",
                "description": "Hamma Post Categorylarni korish",
                "operationId": "Post_category_index",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostCategory"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "PostCategory"
                ],
                "summary": "new Post Category add",
                "description": "Yangi Post Category qoshish",
                "operationId": "Post_Category_store",
                "requestBody": {
                    "description": "save",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name_uz",
                                    "name_ru",
                                    "name_eng"
                                ],
                                "properties": {
                                    "name_uz": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Uzbek"
                                    },
                                    "name_ru": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Russian"
                                    },
                                    "name_eng": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "English"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostCategory"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/post-category/{id}": {
            "get": {
                "tags": [
                    "PostCategory"
                ],
                "summary": "Show category",
                "description": "bitta Post Category hamma malumotlarini korsatadi",
                "operationId": "Post_Cateogry_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostCategory"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "PostCategory"
                ],
                "summary": "Update existing project",
                "description": "Returns updated project data",
                "operationId": "Post_Category_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "PostCategory",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name_uz",
                                    "name_ru",
                                    "name_eng"
                                ],
                                "properties": {
                                    "name_uz": {
                                        "type": "string",
                                        "example": "Harley"
                                    },
                                    "name_ru": {
                                        "type": "string",
                                        "example": "Harley"
                                    },
                                    "name_eng": {
                                        "type": "string",
                                        "example": "Harley"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostCategory"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "PostCategory"
                ],
                "summary": "Delete existing project",
                "description": "Deletes a record and returns no content",
                "operationId": "Post_Category_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostCategory"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PostCategory": {
                "title": "PostCategory",
                "required": [
                    "name_uz",
                    "name_ru",
                    "name_eng"
                ],
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "name_uz": {
                        "type": "string"
                    },
                    "name_ru": {
                        "type": "string"
                    },
                    "name_eng": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}